backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel>=0.0.21
python-dotenv==1.0.0
pytest==7.4.3
pydantic-settings>=2.0.0
PyJWT>=2.8.0
mcp[cli]>=1.26.0
google-generativeai>=0.3.0
